refactor(extensions): split the init scaffold out of extensions-cli
dxkit's own PR gate blocked #130: extensions-cli.ts crossed the
large-file threshold (600 > 500) as list/refresh/dev/init accumulated
across the extension milestone. init shares nothing with the other
subcommands beyond the registries, so it moves whole to
extensions-init-cli.ts (both rungs' scaffolds + stubs) — 429 + 193
lines. Fixed at the finding, not the baseline.
